+++ src/mainwindow.h
@@ -58,6 +58,7 @@ class MainWindow : public GeonkickWidget
       void openExportDialog();
       void openPreset();
       void openAboutDialog();
+      void resetToDefault();
       void updateLimiter(KitModel::PercussionIndex index);
       RK_DECL_ACT(updateGui, updateGui(), RK_ARG_TYPE(), RK_ARG_VAL());